Move more types to seperate file(s)
