+++ src/contexts/GroupsContextDispatch.tsx
@@ -1,31 +1,8 @@
 import { createContext, useReducer } from 'react';
-import { Group, Image } from '../types';
 import groupReducer from '../reducers/group';
-
-export type GroupContextType = {
-  dispatch: React.Dispatch<GroupAction>;
-  state: GroupState;
-};
+import { GroupContextType } from './types';
 
 const GroupsDispatchContext = createContext<GroupContextType | undefined>(undefined);
-
-export type GroupAction = { type: 'setIndex'; payload: number }
-| { type: 'setImages'; payload: Image[] }
-| { type: 'addGroup'; payload: { group: Group; index?: number }}
-| { type: 'replaceGroups'; payload: { groups: Group[]; index: number }} 
-| { type: 'addImage'; payload: Image }
-| { type: 'updateGroupMemberCount'; payload: Group[] }
-| { type: 'init'; payload: {
-  groups: Group[];
-  images: Image[];
-  index: number;
-};};
-
-type GroupState = {
-  groups: Group[];
-  images: Image[];
-  index: number;
-};
 
 type GroupsProviderProps = { children: React.ReactNode };
